+++ app/Services/PlacementTypeService.php
@@ -2,18 +2,21 @@
 
 namespace App\Services;
 
+use Exception;
+use App\Models\Branch;
 use App\Models\PlacementType;
+use Illuminate\Support\Facades\Auth;
 
 class PlacementTypeService extends Service
 {
     public function store(array $data)
     {
-        $placementType = PlacementType::create($data);
-        // $placementType = PlacementType::with([
-        //     'branch',
-        // ])->findOrFail($placementType->id);
+        $branch = Branch::findOrFail($data['branch_id']);
+        if ($branch->client !== Auth::user()->client_id) {
+            throw new Exception("You don't have permission to add a placement type in this branch.", 403);
+        }
 
-        return $placementType;
+        return PlacementType::create($data);
     }
 
     public function update(array $data, mixed $placementType)
@@ -22,6 +25,10 @@ class PlacementTypeService extends Service
 
     public function delete(mixed $placementType)
     {
+        if ($placementType->branch_id !== Auth::user()->client_id) {
+            throw new Exception("You don't have permission to delete this placement type from this branch.", 403);
+        }
+
         $placementType->delete();
     }
 
@@ -33,14 +40,12 @@ class PlacementTypeService extends Service
     {
     }
 
-    public function listByBranch(int $branch, int $perPage = 25, string $query = null)
+    public function listByBranch(int $branch, string $query = null)
     {
-        $placementTypes = PlacementType::where('branch_id', $branch);
-
-        if ($query) {
-            $placementTypes->where('name', 'like', '%'.$query.'%');
-        }
-
-        return $placementTypes->get();
+        return PlacementType::where('branch_id', $branch)
+            ->when($query, function ($q) use ($query) {
+                $q->where('name', 'like', '%'.$query.'%');
+            })
+            ->get();
     }
 }
